third_party/WebKit/LayoutTests/http/tests/locks/lock_attributes.html

<!DOCTYPE html>
<title>Web Locks API: Lock Attributes</title>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script>
'use strict';

promise_test(async t => {
  const lock = await requestLock(['r1', 'r2'], 'shared')
  assert_array_equals(lock.scope, ['r1', 'r2']);
  assert_equals(lock.mode, 'shared');
}, 'Lock attributes reflect requested properties (shared)');

promise_test(async t => {
  const lock = await requestLock(['r1', 'r2'], 'exclusive')
  assert_array_equals(lock.scope, ['r1', 'r2']);
  assert_equals(lock.mode, 'exclusive');
}, 'Lock attributes reflect requested properties (exclusive)');

promise_test(async t => {
  const lock = await requestLock(['r2', 'r1'], 'shared')
  assert_array_equals(lock.scope, ['r1', 'r2']);
}, 'Lock.scope is in sorted order');

</script>
